Remove the detail page and some requirement when creating data template
