+++ webportal/src/app/create_item/components/create_data/data_template.jsx
@@ -48,7 +48,7 @@ const dataTemplate = {
 function generateDataTemplate(dataTemplateObject, itemObject) {
   dataTemplate.prerequisites[0].uri = dataTemplateObject.dockerImage;
   dataTemplate.prerequisites[1].name = dataTemplateObject.storageName;
-  dataTemplate.prerequisites[1].uri = [dataTemplateObject.storageName];
+  dataTemplate.prerequisites[1].uri = [dataTemplateObject.storagePath];
   dataTemplate.taskRoles.taskrole.data = dataTemplateObject.storageName;
   dataTemplate.taskRoles.taskrole.commands = dataTemplateObject.commands.split(
     '\n',
